Fix CUDA OOM on long sequences (true root cause of NCCL timeouts)
The recurring NCCL ALLREDUCE timeouts at SeqNum=52/54 were not a DDP-variance
issue. They were silent CUDA OOMs on whichever rank drew a long-sequence
rollout, which then hung in CUDA while the other ranks waited 30 min at
the next collective.

Two memory bombs in the per-step forward path:

1. gather_token_logprobs / token_entropy used F.log_softmax over [B, T, V],
   materializing a ~20 GiB fp32 tensor at T=16k V=152k. Replaced with a
   chunked gather - logsumexp implementation that processes T in 1024-token
   slices, dropping peak memory to O(B * 1024 * V).

2. policy_forward_with_grad called .float() on raw [B, T, V] bf16 logits,
   adding another ~20 GiB fp32 allocation. Removed; gather_token_logprobs
   now upcasts each chunk internally.

Combined, peak per-rank step memory drops by ~40 GiB on a T=16k batch,
which is what made one rank silently die while the rest blocked on NCCL.

+++ train/train.py
@@ -282,7 +282,10 @@ def policy_forward_with_grad(
             ids = batch.input_ids[s:e]
             attn = batch.attention_mask[s:e]
             amask = batch.assistant_mask[s:e]
-            logits = model(input_ids=ids, attention_mask=attn, use_cache=False).logits.float()
+            # Keep logits in native bf16; gather_token_logprobs/token_entropy
+            # upcast each T-chunk internally. An upfront `.float()` over
+            # [B, T, V] would peak at ~20 GiB at T=16k / V=152k and OOM.
+            logits = model(input_ids=ids, attention_mask=attn, use_cache=False).logits
             out = compute_rl_loss(
                 logits=logits,
                 input_ids=ids,
@@ -305,7 +308,7 @@ def policy_forward_with_grad(
             ids = batch.input_ids[s:e]
             attn = batch.attention_mask[s:e]
             amask = batch.assistant_mask[s:e]
-            logits = model(input_ids=ids, attention_mask=attn, use_cache=False).logits.float()
+            logits = model(input_ids=ids, attention_mask=attn, use_cache=False).logits
             loss = logits.sum() * 0.0
             out = None
             w = 0.0
